backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
pydantic
